+++ packages/nimble-blazor/Tests/NimbleBlazor.Tests/Unit/Components/NimbleTableColumnTextTests.cs
@@ -13,7 +13,7 @@ public class NimbleTableColumnTextTests
     [Fact]
     public void NimbleTableColumnText_WithFieldNameAttribute_HasTableMarkup()
     {
-        var table = RenderWithPropertySet<string>(x => x.FieldName!, "FirstName");
+        var table = RenderWithPropertySet(x => x.FieldName!, "FirstName");
 
         var expectedMarkup = @"field-name=""FirstName""";
         Assert.Contains(expectedMarkup, table.Markup);
@@ -22,7 +22,7 @@ public class NimbleTableColumnTextTests
     [Fact]
     public void NimbleTableColumnText_WithPlaceholderAttribute_HasTableMarkup()
     {
-        var table = RenderWithPropertySet<string>(x => x.Placeholder!, "No Value");
+        var table = RenderWithPropertySet(x => x.Placeholder!, "No Value");
 
         var expectedMarkup = @"placeholder=""No Value""";
         Assert.Contains(expectedMarkup, table.Markup);
@@ -31,7 +31,7 @@ public class NimbleTableColumnTextTests
     [Fact]
     public void NimbleTableColumnText_WithFractionalWidthAttribute_HasTableMarkup()
     {
-        var table = RenderWithPropertySet<double>(x => x.FractionalWidth!, 2);
+        var table = RenderWithPropertySet(x => x.FractionalWidth!, 2);
 
         var expectedMarkup = @"fractional-width=""2""";
         Assert.Contains(expectedMarkup, table.Markup);
@@ -40,9 +40,45 @@ public class NimbleTableColumnTextTests
     [Fact]
     public void NimbleTableColumnText_WithMinPixelWidthAttribute_HasTableMarkup()
     {
-        var table = RenderWithPropertySet<double?>(x => x.MinPixelWidth!, 40);
+        var table = RenderWithPropertySet(x => x.MinPixelWidth!, 40);
 
         var expectedMarkup = @"min-pixel-width=""40""";
+        Assert.Contains(expectedMarkup, table.Markup);
+    }
+
+    [Fact]
+    public void NimbleTableColumnText_WithSortIndexAttribute_HasTableMarkup()
+    {
+        var table = RenderWithPropertySet(x => x.SortIndex!, 0);
+
+        var expectedMarkup = @"sort-index=""0""";
+        Assert.Contains(expectedMarkup, table.Markup);
+    }
+
+    [Fact]
+    public void NimbleTableColumnText_WithSortDirectionAttribute_HasTableMarkup()
+    {
+        var table = RenderWithPropertySet(x => x.SortDirection!, TableColumnSortDirection.Descending);
+
+        var expectedMarkup = @"sort-direction=""descending""";
+        Assert.Contains(expectedMarkup, table.Markup);
+    }
+
+    [Fact]
+    public void NimbleTableColumnText_WithGroupIndexAttribute_HasTableMarkup()
+    {
+        var table = RenderWithPropertySet(x => x.GroupIndex!, 0);
+
+        var expectedMarkup = @"group-index=""0""";
+        Assert.Contains(expectedMarkup, table.Markup);
+    }
+
+    [Fact]
+    public void NimbleTableColumnText_WithGroupingDisabledAttribute_HasTableMarkup()
+    {
+        var table = RenderWithPropertySet(x => x.GroupingDisabled!, true);
+
+        var expectedMarkup = @"grouping-disabled";
         Assert.Contains(expectedMarkup, table.Markup);
     }