+++ packages/nimble-components/.storybook/main.js
@@ -1,40 +1,53 @@
-const CircularDependencyPlugin = require('circular-dependency-plugin');
+import remarkGfm from 'remark-gfm';
+import CircularDependencyPlugin from 'circular-dependency-plugin';
 
-module.exports = {
-    stories: ['../src/**/*.stories.@(ts|mdx)', '../docs/**/*.stories.mdx'],
-    addons: [
-        {
-            name: '@storybook/addon-essentials',
-            options: {
-                outline: false
-            }
-        },
-        '@storybook/addon-a11y',
-        '@storybook/addon-interactions'
-    ],
-    webpackFinal: config => {
-        config.module.rules.push({
-            test: /\.ts$/,
-            use: [
-                {
-                    loader: require.resolve('ts-loader')
-                }
-            ]
-        });
-        config.plugins.push(
-            new CircularDependencyPlugin({
-                exclude: /node_modules/,
-                failOnError: process.env.NODE_ENV === 'production'
-            })
-        );
-        config.performance = {
-            hints: false
-        };
-        return config;
+export const stories = [
+    '../docs/**/*.mdx',
+    '../src/**/*.mdx',
+    '../src/**/*.stories.ts'
+];
+export const addons = [
+    {
+        name: '@storybook/addon-essentials',
+        options: {
+            outline: false
+        }
     },
-    staticDirs: ['public'],
-    framework: {
-        name: '@storybook/html-webpack5',
-        options: {}
-    }
+    {
+        name: '@storybook/addon-docs',
+        options: {
+            mdxPluginOptions: {
+                mdxCompileOptions: {
+                    remarkPlugins: [remarkGfm]
+                }
+            }
+        }
+    },
+    '@storybook/addon-a11y',
+    '@storybook/addon-interactions'
+];
+export function webpackFinal(config) {
+    config.module.rules.push({
+        test: /\.ts$/,
+        use: [
+            {
+                loader: require.resolve('ts-loader')
+            }
+        ]
+    });
+    config.plugins.push(
+        new CircularDependencyPlugin({
+            exclude: /node_modules/,
+            failOnError: process.env.NODE_ENV === 'production'
+        })
+    );
+    config.performance = {
+        hints: false
+    };
+    return config;
+}
+export const staticDirs = ['public'];
+export const framework = {
+    name: '@storybook/html-webpack5',
+    options: {}
 };