+++ packages/nimble-components/src/table/components/row/tests/table-row.spec.ts
@@ -50,7 +50,7 @@ describe('TableRow', () => {
         const renderedCell = pageObject.getRenderedCell(0);
 
         expect(renderedCell!.cellViewTemplate).toEqual(
-            textColumn1.currentCellViewTemplate
+            textColumn1.columnInternals.cellViewTemplate
         );
     });
 
@@ -81,5 +81,40 @@ describe('TableRow', () => {
         const secondCellRecord = secondCellState!
             .cellRecord as TableColumnTextCellRecord;
         expect(secondCellRecord.value).toBe('foo');
+    });
+
+    it('does not have aria-selected attribute when it is not selectable', async () => {
+        element.selectable = false;
+        element.selected = false;
+        await connect();
+
+        expect(element.hasAttribute('aria-selected')).toBeFalse();
+    });
+
+    it('has aria-selected attribute set to "true" when it is selected', async () => {
+        element.selectable = true;
+        element.selected = true;
+        await connect();
+
+        expect(element.getAttribute('aria-selected')).toBe('true');
+    });
+
+    it('has aria-selected attribute set to "false" when it is not selected', async () => {
+        element.selectable = true;
+        element.selected = false;
+        await connect();
+
+        expect(element.getAttribute('aria-selected')).toBe('false');
+    });
+
+    it('updates aria-selected attribute when selection state changes', async () => {
+        element.selectable = true;
+        element.selected = false;
+        await connect();
+
+        element.selected = true;
+        await waitForUpdatesAsync();
+
+        expect(element.getAttribute('aria-selected')).toBe('true');
     });
 });